Fix scroll to text WPT flakiness
Since stash.py polls until results are available, the tests can time out
in a single threaded environment if the test page tries to fetch results
before the target page stashes them. This patch moves the polling to
stash.js instead, so that the target page is free to stash results.

Bug: https://github.com/web-platform-tests/wpt/issues/20269
Change-Id: Ia2df8f889596e91207aca92f6ecb290c9f96d201