Use assert_precondition on LayoutInstability and EventTiming
This CL replaces some assert_unreached with assert_precondition and adds
some assert_precondition to tests. The ones that use ScoreWatcher should
already throw an error and not timeout. async_test should not use async
since that will result in errors not being caught, so this CL fixes that
too.

Change-Id: Iecda47acce998efac0dc78e5214b60b005c16075